// src/lib/shared-models/models/UserActivity.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NCorp.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { UserActivityType } from './UserActivityType';
import {
    UserActivityTypeFromJSON,
    UserActivityTypeFromJSONTyped,
    UserActivityTypeToJSON,
    UserActivityTypeToJSONTyped,
} from './UserActivityType';

/**
 * 
 * @export
 * @interface UserActivity
 */
export interface UserActivity {
    /**
     * 
     * @type {UserActivityType}
     * @memberof UserActivity
     */
    activityType?: UserActivityType;
    /**
     * 
     * @type {string}
     * @memberof UserActivity
     */
    gameSlug?: string | null;
    /**
     * 
     * @type {string}
     * @memberof UserActivity
     */
    gameTitle?: string | null;
}



/**
 * Check if a given object implements the UserActivity interface.
 */
export function instanceOfUserActivity(value: object): value is UserActivity {
    return true;
}

export function UserActivityFromJSON(json: any): UserActivity {
    return UserActivityFromJSONTyped(json, false);
}

export function UserActivityFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserActivity {
    if (json == null) {
        return json;
    }
    return {
        
        'activityType': json['activityType'] == null ? undefined : UserActivityTypeFromJSON(json['activityType']),
        'gameSlug': json['gameSlug'] == null ? undefined : json['gameSlug'],
        'gameTitle': json['gameTitle'] == null ? undefined : json['gameTitle'],
    };
}

export function UserActivityToJSON(json: any): UserActivity {
    return UserActivityToJSONTyped(json, false);
}

export function UserActivityToJSONTyped(value?: UserActivity | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'activityType': UserActivityTypeToJSON(value['activityType']),
        'gameSlug': value['gameSlug'],
        'gameTitle': value['gameTitle'],
    };
}