// src/lib/shared-models/models/UserConnectionStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * NCorp.Server
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * 
 * @export
 */
export const UserConnectionStatus = {
    Connected: 'Connected',
    Idle: 'Idle',
    Disconnected: 'Disconnected'
} as const;
export type UserConnectionStatus = typeof UserConnectionStatus[keyof typeof UserConnectionStatus];


export function instanceOfUserConnectionStatus(value: any): boolean {
    for (const key in UserConnectionStatus) {
        if (Object.prototype.hasOwnProperty.call(UserConnectionStatus, key)) {
            if (UserConnectionStatus[key as keyof typeof UserConnectionStatus] === value) {
                return true;
            }
        }
    }
    return false;
}

export function UserConnectionStatusFromJSON(json: any): UserConnectionStatus {
    return UserConnectionStatusFromJSONTyped(json, false);
}

export function UserConnectionStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): UserConnectionStatus {
    return json as UserConnectionStatus;
}

export function UserConnectionStatusToJSON(value?: UserConnectionStatus | null): any {
    return value as any;
}

export function UserConnectionStatusToJSONTyped(value: any, ignoreDiscriminator: boolean): UserConnectionStatus {
    return value as UserConnectionStatus;
}